[lld-macho] Add initial support for chained fixups
This commit adds support for chained fixups, which were introduced in
Apple's late 2020 OS releases. This format replaces the dyld opcodes
used for supplying rebase and binding information. A high-level overview
of the format and my implementation is found in SyntheticSections.h.

This feature is not enabled by default for now, as further real world
testing is required to ensure there are no regressions.

Differential Revision: https://reviews.llvm.org/D132560